fix(code-editor): fix hooks' arguments types
